refactor: relocate listShops util
