+++ apps/cms/src/app/cms/dashboard/page.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { listShops } from "../listShops";
+import { listShops } from "../../../lib/listShops";
 
 export default async function DashboardIndexPage() {
   const shops = await listShops();
