+++ apps/cms/src/app/cms/settings/page.tsx
@@ -1,6 +1,6 @@
 // apps/cms/src/app/cms/settings/page.tsx
 import Link from "next/link";
-import { listShops } from "../listShops";
+import { listShops } from "../../../lib/listShops";
 
 export default async function SettingsIndexPage() {
   const shops = await listShops();
