+++ apps/cms/src/app/cms/shop/page.tsx
@@ -1,7 +1,7 @@
 // apps/cms/src/app/cms/shop/page.tsx
 
 import Link from "next/link";
-import { listShops } from "../listShops";
+import { listShops } from "../../../lib/listShops";
 
 export const metadata = {
   title: "Choose shop · Base-Shop",
